refractor to have itemsCounter separate function from populatePage
